need to fix all the prints + new putnbr functions

+++ printf_git/ft_print_hex.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_print_hex.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: btoksoez <btoksoez@student.42.fr>          +#+  +:+       +#+        */
+/*   By: btoksoez <btoksoez@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 10:46:44 by btoksoez          #+#    #+#             */
-/*   Updated: 2023/10/21 18:02:21 by btoksoez         ###   ########.fr       */
+/*   Updated: 2023/10/24 14:12:38 by btoksoez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 static void	ft_print_hex_recursive(void *p)
 {
-	char		c;
-	uintptr_t	int_p;
+	char			c;
+	unsigned long	int_p;
 
-	int_p = (uintptr_t)p;
+	int_p = (unsigned long)p;
 	if (int_p >= 16)
 		ft_print_hex_recursive((void *)(int_p / 16));
 	c = "0123456789abcdef"[int_p % 16];
@@ -30,13 +30,10 @@ void	ft_print_hex(void *p)
 	ft_print_hex_recursive(p);
 }
 // #include <stdio.h>
-
 // int main(void)
 // {
 // 	void *p;
-// 	char c = 'c';
-
-// 	p = &c;
+// 	p = (void *)0;
 // 	ft_print_hex(p);
 // 	write(1, "\n", 1);
 // 	printf("%p", p);
